refactor recommend app
