+++ backend/vesta/kitchen/urls.py
@@ -19,5 +19,4 @@ urlpatterns = [
     path('menu/<str:menuname>/', views.menu_name, name='menuName'),
     path('token/', views.token, name='token'),
     # path('ml/detection/', views.detection, name='detection'),
-    path('recommend/<str:date>/', views.recommend, name='recommend'),
 ]
