+++ backend/vesta/recommend/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('<str:date>/', views.recommend, name='recommend'),
+]
