+++ backend/vesta/vesta/urls.py
@@ -20,5 +20,6 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+    path('api/recommend/', include('recommend.urls')),
     path('api/', include('kitchen.urls')),
 ]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
